box-upload script - determine provider from dist

+++ scripts/box-upload.py
@@ -6,17 +6,17 @@ You need to specify device ip address in ./secret.json file.
 from subprocess import check_call
 import os
 import json
+from src.dist import NAME
 
 if __name__ == "__main__":
 
-	provider = "all"
 	build_dir = "/tmp/iptvdream-build"
 
 	check_call(["docker", "exec", "-it", "enigma2", "rm", "-rf", build_dir])
 
 	make = [
 		"docker", "exec", "-it",
-		"-e", "PROVIDER=%s" % provider,
+		"-e", "PROVIDER=%s" % NAME,
 		"-e", "DESTDIR=%s" % build_dir,
 		"enigma2", "make"
 	]
@@ -27,7 +27,8 @@ if __name__ == "__main__":
 		data = json.load(f)
 		ipk = "packages/%s_%s_all.ipk" % (data['name'], data['version'])
 
-	os.remove(ipk)
+	if os.path.isfile(ipk):
+		os.remove(ipk)
 	check_call(make + [ipk])
 	assert os.path.isfile(ipk)
